chore(types): split account and account-with-password types
